backend/requirements.txt
fastapi
uvicorn
websockets
pydantic
pydantic-settings
python-jose
passlib
bcrypt
python-multipart
sqlalchemy
pymysql
redis
celery
aiosmtplib
twilio
python-dotenv
boto3
cryptography
email-validator
pytest
httpx